+++ src/main/java/com/fcgl/madrid/forum/model/request/PostRequest.java
@@ -5,7 +5,7 @@ public class PostRequest {
     private String title;
     private String description;
     private Integer cityId;
-    private Integer userId;
+    private Long userId;
 
     /**
      * Request Parameters for post
@@ -14,7 +14,7 @@ public class PostRequest {
      * @param cityId
      * @param userId
      */
-    public PostRequest(String title, String description, Integer cityId, Integer userId) {
+    public PostRequest(String title, String description, Integer cityId, Long userId) {
         this.cityId = cityId;
         this.title = title;
         this.description = description;
@@ -45,11 +45,11 @@ public class PostRequest {
         this.cityId = cityId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
